# swagger/swagger.yaml
swagger: "2.0"
info:
  title: Movie Character API
  description: Movie Character API description
  version: 1.0.0
host: "localhost:8080"
basePath: "/api/v1"
paths:
...
  /franchise/{id}:
      get:
        description: Get the Franchise with the given id
        produces:
        - application/json
        parameters:
        - name: "id"
          in: "path"
          description: "ID of Franchise to return"
          required: true
          type: "integer"
        responses:
          '200':
            description: The Franchise with the given id.
            schema:
              $ref: '#/definitions/ReturnFranchise'
          '400':
            description: No franchise with the given id.
      put:
        description: Update the Franchise with the given id.
        consumes:
        - "application/json"
        produces:
        - application/json
        parameters:
        - name: "id"
          in: "path"
          description: "ID of Franchise to update"
          required: true
          type: "integer"
        - in: "body"
          name: "body"
          description: "Franchise object that needs to be added to the store"
          required: true
          schema:
            $ref: "#/definitions/Franchise"
        responses:
          '200':
            description: The Franchise with the given id.
            schema:
              $ref: '#/definitions/MovieCharacter'
      delete:
        description: Delete the Franchise with the given id.
        parameters:
          - name: "id"
            in: "path"
            description: "ID of Franchise to delete"
            required: true
            type: "integer"
        responses:
          '204':
            description: Franchise deleted successfully
  /franchise/{id}/characters:
    get:
      description: Get all characters in the given franchise
      produces:
        - application/json
      parameters:
        - name: "id"
          in: "path"
          description: "ID of Franchise to return"
          required: true
          type: "integer"
      responses:
        '200':
          description: The Franchise with the given id.
          schema:
            type: "array"
            items:
              $ref: '#/definitions/ReturnMovieCharacter'
        '400':
          description: No franchise with the given id.
  /franchise/{id}/movies:
    get:
      description: Get all movies in the given franchise
      produces:
        - application/json
      parameters:
        - name: "id"
          in: "path"
          description: "ID of franchise to return"
          required: true
          type: "integer"
      responses:
        '200':
          description: The Franchise with the given id.
          schema:
            type: "array"
            items:
              $ref: '#/definitions/ReturnMovie'
        '400':
          description: No franchise with the given id.
...
definitions:
  Franchise:
    type: "object"
    properties:
      id:
        type: "integer"
      name:
        type: "string"
      description:
        type: "string"
      movies:
        type: "array"
        items:
          type: "object"
          properties:
            id:
              type: integer
  ReturnFranchise:
    type: "object"
    properties:
      id:
        type: "integer"
      name:
        type: "string"
      description:
        type: "string"
      movies:
        type: "array"
        items:
          type: "string"
...
  ReturnMovie:
    type: "object"
    properties:
      id:
        type: "integer"
      title:
        type: "string"
      genre:
        type: "string"
      release:
        type: "string"
        format: "date"
      director:
        type: "string"
      picture:
        type: "string"
      trailer:
        type: "string"
      franchise:
        type: "string"
      movieCharacters:
        type: "array"
        items:
          type: "string"
...
  MovieCharacter:
    type: "object"
    properties:
      id:
        type: "integer"
      firstName:
        type: "string"
      lastName:
        type: "string"
      alias:
        type: "string"
      gender:
        type: "string"
        enum: [FEMALE, MALE, OTHER]
      picture:
        type: "string"
      movies:
        type: "array"
        items:
          type: "object"
          properties:
            id:
              type: integer
  ReturnMovieCharacter:
    type: "object"
    properties:
      id:
        type: "integer"
      firstName:
        type: "string"
      lastName:
        type: "string"
      alias:
        type: "string"
      gender:
        type: "string"
        enum: [FEMALE, MALE, OTHER]
      picture:
        type: "string"
      movies:
        type: "array"
        items: 
          type: "string"
...